requirements.txt
Flask==3.0.0
Flask_Login==0.6.3
flask_wtf==1.2.1
kafka_python==2.0.2
pymongo==4.6.1
Werkzeug==3.0.1
WTForms==3.1.1
